Added replaces_plan_uuid to API and database
This is a need Luke identified to capture a proposed plan's intent -- where, if for example an actor is attempting to become the primary in a relationship, they can't do this by overwriting an existing relationship unless both actors are really clear that that is what's happening. This means we have to say "this is the plan I want to replace" so the other side can say "wait, that's not okay with me, I was doing business with you under different conditions". Without this uuid reference, they may think that this is an additional agreement not intended to supercede the old.