Remove PUT method for grains and add to POST method
Some feedback we got is that the distinction between PUT and POST is too esoteric and implemented haphazardly by all. Therefore it's better to accept some of the ambiguity and standardize on one method, POST.

This means one substantive change to the meaning of the response code 409. It now will be given if the grain exists elsewhere AND the payload was supplied. Previously the payload would be discarded or ignored and only the UUID linked, because this was a separate method and could be deterministically set up safely without requring checking all the content of given grains to see if they are different.